Train and load a dense DoRA on three architectures
Phase 3. Z-Image, Lens and MiniT2I take ("lora", True) on both axes; DoHa
and DoKr stay refused everywhere, because the engine builds them but no
architecture has a round trip for them yet.

SD1.5 and SDXL are refused on measurement rather than on the no-session
rule. Diffusers drops every dora_scale key with a log warning before its
Kohya converter runs, so the branch that would map one onto a magnitude
vector is dead on that path: a DoRA applied to 40 of 40 U-Net targets,
installed no magnitude, and came out byte-identical to the same file with
its magnitudes stripped -- a successful generation at the wrong numbers.
Until this commit nothing here could write such a file; now three
architectures can, into the same directory the SD1.5 selector lists from,
so load_loras refuses a weight-decomposed file by name. LoHa and LoKr on
that path already fail loudly inside the converter, so they keep doing that
and a test records it.

A weight-only quantized base is refused at four points, keyed on the
weight's dtype rather than on the quantized Linear classes, because the
legacy FP8 pass leaves an ordinary Linear holding a float8 weight. The
fourth point is one no install-time check could see: Z-Image loads adapters
sixty lines before it quantizes, and that pass deep-copies and casts every
Linear weight including the wrapper's own base, which moves the delta 1.3%
with no error. The quantization drops instead, and the guard is a no-op for
the additive families.

The magnitude reaches the optimizer exactly once under both fused paths, is
written per target, and a resume without it refuses by name rather than
restarting it at the base's row norms. dora_scale now takes the branch's
dtype rather than the base's, so an fp16 run no longer trains it without an
fp32 master while its factors keep one.

Strength-zero identity is gated against a perturbed magnitude: a fresh
layer's magnitude equals the base's row norms, which makes s=0 an identity
by coincidence. PEFT's own DoRA layer is a third-party witness for the row
convention at 1e-7; the column form still has none, and the oracle says so.

+++ backend/core/adapters/groups.py
@@ -395,7 +395,10 @@ def build_adapter_branch(
         branch = cls.from_tensors(base, group, alpha=alpha, lora_dtype=lora_dtype,
                                   lora_name=lora_name or group.stem)
         if group.weight_decompose:
-            branch = DoRALinearLayer(base, branch, dora_scale=group["dora_scale"])
+            # Same dtype as the factors it rides with -- usually the base
+            # weight's own, though Lens's branch dtype prefers the bias.
+            branch = DoRALinearLayer(base, branch, dora_scale=group["dora_scale"],
+                                     dtype=lora_dtype)
     except (IndexError, KeyError, RuntimeError, TypeError, ValueError,
             ZeroDivisionError):
         return SHAPE_MISMATCH
